lecture_summary_final/lecture_summary7/lecture_summary7/lecture_summary/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-cloud-speech==2.33.0
google-cloud-storage==3.4.0
pydub==0.25.1
python-dotenv==1.0.0
google-generativeai==0.8.3
google-genai==0.2.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.182.0
python-docx==1.1.0
reportlab==4.0.7
python-bidi==0.4.2
arabic-reshaper==3.0.0
